pyproject.toml
[tool.poetry]
name = "livekit-agent"
version = "0.1.0"
description = ""
authors = ["smartinezbragado <smartinezbragado@gmail.com>"]
readme = "README.md"
packages = [{include = "livekit_agent"}]

[tool.poetry.dependencies]
python = "^3.12"
livekit-agents = "^0.11.1"
livekit-plugins-silero = "^0.7.3"
python-dotenv = "^1.0.1"
livekit-plugins-openai = "^0.10.5"
livekit-plugins-elevenlabs = "^0.7.7"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
click = "^8.1.7"
loguru = "^0.7.2"
requests = "^2.32.3"
sounddevice = "^0.5.1"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
